refactor(db): Move statistics submission into dedicated class
* Add DbStatsSubmit class
